Fix: Handle empty input correctly & change int to double

+++ src/lv2/Calculator.java
@@ -4,10 +4,10 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Calculator {
-    private List<Integer> results =  new ArrayList<Integer>();
+    private List<Double> results =  new ArrayList<Double>();
 
-    public Integer calculate(int num1, int num2, char operator){
-        int result = 0;
+    public Double calculate(int num1, int num2, char operator){
+        double result = 0;
         switch (operator) {
             case '+':
                 result = num1 + num2;
@@ -24,7 +24,7 @@ public class Calculator {
                     System.out.println("나눗셈 연산에서 분모(두 번째 정수)에 0이 입력될 수 없습니다.");
                     return null;
                 }
-                result = num1 / num2;
+                result = (double) num1 / num2;
                 break;
             default:
                 System.out.println("잘못된 입력입니다.");
@@ -34,11 +34,11 @@ public class Calculator {
         return result;
     }
 
-    public List<Integer> getResults() {
+    public List<Double> getResults() {
         return results;
     }
 
-    public void setResults(List<Integer> results) {
+    public void setResults(List<Double> results) {
         this.results = results;
     }
